requirements.txt
discord==1.7.3
discord.py==1.7.3
parsedatetime>=2.6
dateparser>=1.0.0
python-dotenv
googlesearch-python==1.1.0
lxml
yt-dlp==2022.9.1
flask
